saige/requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0

# Google Gemini Integration
langchain-google-genai>=2.0.0
google-auth>=2.0.0  # Required for Google Cloud service account authentication

# Vertex AI Python SDK for managing indexes and endpoints
# Pinned to 1.138.0 for compatibility with google-api-core 2.30.0 and protobuf 6.x
google-cloud-aiplatform==1.138.0

# Data Validation & Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi
uvicorn[standard]>=0.27.0
langchain-google-genai

# Google Cloud Firestore (for RAG system)
google-cloud-firestore>=2.0.0

#sql
pymssql

# Weather API
requests>=2.31.0

# Redis (for checkpointing and message buffer)
redis>=5.0.0
langgraph-checkpoint-redis>=0.3.0 # python version for venv >3.10 or <3.14 

# Testing (Task 6)
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0